Trimming adapters and low quality base calls
Trimming is an important step when the reads sequences have adapters not removed and low-quality regions. Trimmomatic is a commonly used tool for this purpose, and in this code, we use it to remove the Illumina adapters and low quality base calls. This shell script will produce four output files for each sample.